requirements.txt
streamlit
PyPDF2
sumy
nltk
transformers
torch
python-docx
